python-project/requirements.txt
# Dependências principais do projeto
google-cloud-documentai>=2.20.0
python-dotenv>=1.0.0
